FIX
Fixed so that the development is added to Imperial Palace to get the development of the Emperor to the development of the 9th Great Powers OWN development. Must add changes to add on 9th GP subjects' development to this sum as well as it is used to calculate the GP rank.